requirements.txt
# Core processing
opencv-python>=4.8.0
numpy>=1.21.0
pillow>=9.0.0
pyyaml>=6.0
tqdm>=4.64.0
python-dotenv>=1.0.0

# Machine learning and diffusion models
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
peft>=0.4.0

# Computer vision and image processing
scikit-image>=0.19.0

# External APIs
boto3>=1.26.0
botocore>=1.29.0
requests>=2.28.0

# Image hashing for deduplication
imagehash>=4.3.0

# Optional: Alternative caption providers
# openai>=1.0.0
# anthropic>=0.3.0

# Development (optional)
# pytest>=7.0.0
# black>=22.0.0
# flake8>=5.0.0
